+++ FirebaseStorage/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+- [fixed] `putFile` and `putFileAsync` now work in app extensions. A background session
+   configuration is not used when uploading from an app extension (#12579).
+
 # 10.11.0
 - [added] Add progress tracking capability for `putDataAsync`, `putFileAsync`, and
   `writeAsync`. (#10574)